Emit `bindgen_original_name` attributes for nested types.
This is needed to allow autocxx to use the correct C++ name of the
nested type in generated code. See
https://github.com/google/autocxx/issues/115 for details.

I've updated some tests with the expected attributes. Hopefully, I've
caught all tests that need updating; not all tests currently pass, and I
didn't want to update the expected output for some of the tests that
don't pass because I'm not sure if it's correct to do so.