Allow specifying opaque types precisely.
Previously, opaque types could be specified only using a regex on the
Options type or command line.

This worked well for:
* everything opaque
* a few specific types opaque

but it didn't work for the case where callers want everything opaque,
_except_ a specific few types.

This change adds a callback to the ParseCallbacks trait such that callers
can pick a specific 'opacity' for individual types.