Redo annotations for C++ semantic attributes.
autocxx_bindgen previously added all sorts of arbitrary attributes to
various items, conveying metadata about their meaning in C++ (on top of
the basic information conveyed by their representation in Rust).

Example: whether a parameter is a pointer or a reference, which is
represented the same in the ABI used to link Rust and C++, but conveys
extra meaning in C++ (typically references cannot be null).

This change regularizes such annotations and makes them switchable
using a command line parameter. It is a step towards conceivably,
one day, being able to propose this as an upstream change to start
to unfork autocxx_bindgen.