Fix u16 enum repr.
A previous change assumed we could always treat u16 as char16_t, since
it appeared only to be used to represent char16_t. It turns out it's
also used for the representation of 16-bit-wide enums, so we need to
keep track of the distinction between these two types as they travel
through autocxx-bindgen.

Relates to https://github.com/google/autocxx/issues/1214